move stage particle boundaries to compile unit
ci: picongpu
